requirements.txt
pytest
openpyxl==3.0.0
pandas==0.25.1
jira==2.0.0
atlassian-python-api==1.14.1